refactor(#792): extract block_yaml load/parse helpers into their own file (sweep)
block_yaml.rs 671 -> 500 (leaves the debt list; also drops stale nam/processor
entry). Moved load_audio_block_value / load_model_params / migrate_legacy_model_id
/ extract_core_block_fields to block_yaml_load.rs; the AudioBlockYaml types + impl
stay. Callers (chain_yaml, lib) updated. infra-yaml tests (84) pass, zero warnings.
