+++ crates/infra-yaml/src/lib.rs
@@ -20,9 +20,11 @@ use std::fs;
 use std::path::{Path, PathBuf};
 
 mod block_yaml;
+mod block_yaml_load;
 mod chain_yaml;
 mod rig_yaml;
-use block_yaml::{load_audio_block_value, AudioBlockYaml};
+use block_yaml::AudioBlockYaml;
+use block_yaml_load::load_audio_block_value;
 use chain_yaml::ChainYaml;
 pub use rig_yaml::{
     load_project_any, load_rig_project_file, migrate_legacy_project_file, parse_rig_project,
